Update ROOMMATE APP BETA.py
Learnt the importance of creating a db connection for each function and closing it. Without the code

db_connection = mysql.connector.connect(
        host="localhost",
        user="your_mysql_username",
        password="your_mysql_password",
        database="roommate_app"
    )

    cur = db_connection.cursor()
    
... the helper-function get_student_id_by_gender(gender) would only let me call* get_student_id_by_bedtime_preference()
once - the second call and any call after that would result in an -out-of-sync- db error. I used print statements and
error log checking to link the problem to this helper function and the fact that it was using the global db-connection
because it did not have its own.

On adding it, I can now fetch student ids and full names specific to a bedtime-preference, based on the data in the
database. Example python console outputs for selecting dropdown times 22:00 and 23:00:

bedtime preference: 22
gender: F
Students with bedtime 22:00

Student_id | Name


107 | Laura Moore
113 | Chloe Harris
119 | Mia Lewis
125 | Abigail Scott
131 | Ella Campbell
137 | Emily Evans
127.0.0.1 - - [02/Nov/2024 13:52:45] "POST /get_student_id_by_bedtime_preference HTTP/1.1" 200 -
127.0.0.1 - - [02/Nov/2024 13:52:49] "OPTIONS /get_student_id_by_bedtime_preference HTTP/1.1" 200 -
bedtime preference: 23
gender: F
Students with bedtime 23:00

Student_id | Name


103 | Emily Johnson
109 | Sophia Anderson
115 | Isabella Thompson
121 | Ella Hall
127 | Grace Baker
133 | Amelia Collins
139 | Mila Rivera
140 | Lily Gonzalez

* -> called by pressing the button on the html page